Rename smlbench2 to data-transfer-datasets-random
Still fails at runtime.
